Updated logistic_regression
Started decomposing the problem with the numpy arrays with a simple example. Created train_LogisticRegression.py and made radical changes to the implementation as the data was understood better. There are several problems using numpy that need to be fixed. It is critical to find what's wrong with the dw.